third_party/blink/web_tests/fast/css/image-orientation/image-orientation-expected.txt

Apply image-orientation property and check computed style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS img.style.cssText is "image-orientation: from-image;"
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is "image-orientation: none;"
PASS getComputedStyle(img).imageOrientation is "none"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS successfullyParsed is true

TEST COMPLETE
